refactor(Util) : 스트링을 받아 ','로 split하여 list로 반환하는 기능을 Util로 옮기고 그에 따른 전체 코드 리펙토링
