+++ src/main/java/racingcar/policy/CarMoveForward.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.policy;
 
 public interface CarMoveForward {
     boolean isCarMoveForward(int randomNumber);
