+++ src/main/java/racingcar/policy/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.policy;
 
 public interface NumberGenerator {
     int getRandomDecimalNumber();
