+++ src/main/java/racingcar/util/Util.java
@@ -1,5 +1,8 @@
 package racingcar.util;
 
+import java.util.Arrays;
+import java.util.List;
+
 public class Util {
     public boolean isNullOrEmpty(String string) {
         if (string == null) {
@@ -12,5 +15,10 @@ public class Util {
             return true;
         }
         return false;
+    }
+
+    public List<String> splitByComma(String string) {
+        String[] carNameArray = string.split(",", -1);
+        return Arrays.asList(carNameArray);
     }
 }
